requirements_production.txt
# Production API Dependencies (Minimal Setup)

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0

# API Clients (for main2.py)
openai>=1.0.0
anthropic>=0.7.0
httpx>=0.25.0
requests>=2.31.0

# Basic Utils
python-dotenv>=1.0.0

# Optional: Production Server
# gunicorn>=21.2.0

# Optional: PostgreSQL driver (uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.0

# ============================================
# IMPORTANT: If main2.py needs these, install separately:
# ============================================
# transformers
# torch
# sentence-transformers
# numpy
# pandas
